% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder_lm_plot.R
\name{gapminder_lm_plot}
\alias{gapminder_lm_plot}
\title{Draw a plot to show how well pop/lifeExp/gdpPercap and year in gapminder dataset fits a linear model}
\usage{
gapminder_lm_plot(countryName, x = "year", y = "pop")
}
\arguments{
\item{countryName}{character this should be the name of country that you want to explore in gapminder dataset.}

\item{x}{character default value would be "year" here.}

\item{y}{character this could be one of the three variables(pop/lifeExp/gdpPercap) in gapminder dataset. Default value of y is "pop".}
}
\description{
gapminder_lm_plot is used to show how well pop/lifeExp/gdpPercap and year of a specific country in gapminder dataset
fit the linear model. After using linear model to fit the data, it will display how actual value of pop/lifeExp/gdpPercap changes
with year, as well as how predicted value of pop/lifeExp/gdpPercap changes with year.
}
\examples{
gapminder_lm_plot("Canada")
gapminder_lm_plot("Canada","year","lifeExp")
}

