% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder_lmfit.R
\name{gapminder_lmfit}
\alias{gapminder_lmfit}
\title{Model Population VS Year for countries in gapminder dataset.}
\usage{
gapminder_lmfit(data, countryName, offset = 1952)
}
\arguments{
\item{data}{Here would be gapminder.}

\item{countryName}{character this should be the name of country that you want to explore in gapminder dataset.}

\item{offset}{this can be used to specify an a priori known component to be included in the linear predictor
during fitting. In this function, the default value of offset is 1952, because we think it makes more sense for
the intercept to correspond to population in 1952, the earliest date in gapminder dataset.}
}
\value{
list
}
\description{
gapminder_lmfit is used to fit linear models to population and time for a specific country in gapminder dataset.
It can be used to carry out regression between population and time, then it will return the coefficients, residuals
and fitError(=sum of residuals^2) of fitted linear model back at the same time.
}
\examples{
gapminder_lmfit(gapminder,"Canada")

}

