% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder_lmfit.R
\name{gapminder_lmfit}
\alias{gapminder_lmfit}
\title{Model Population VS Year for countries in GAPMINDER dataset.}
\usage{
gapminder_lmfit(data, offset = 1952)
}
\arguments{
\item{data}{a data frame, containing the variables "population" and "year" for a country in GAPMINDER dataset.}

\item{offset}{this can be used to specify an a priori known component to be included in the linear predictor
during fitting. In this function, the default value of offset is 1952, because we want to make the intersect
of fitted linear model be the predicted value for year 1952, the year when GAPMINDER dataset began to record
data.}
}
\value{
list
}
\description{
gapminder_lmfit is used to fit linear models to population and time for countries in GAPMINDER dataset.
It can be used to carry out regression between population and time, and return the coefficients, residuals
and fitError(=sum of residuals^2) back at the same time.
}
\examples{
canadaData <- gapminder \%>\% filter(country=="Canada") \%>\% arrange(year) \%>\% select(year,pop)
gapminder_lmfit(canadaData)

}

